Add Gitpan::Config to deal with finding and reading the config file.
I'm not 100% sure how overlay merging is going to behave with anything complicated,
but we'll deal with that when it comes up.

Right now the only thing we need is Github logins for testing and production.
Seemed like it made more sense to have the overlays separate than to have "test"
and "live" versions of every key.  We'll see how that turns out.