Fix delete_all_repos.
Gitpan::Github->new needs a repo name, we can use a placeholder since
we're using it as a generic Github interface.

This is probably going to be a problem.  The repo name is for a lot of
things, but is required for the CanDistLog role.  This suggests
Gitpan::Github needs to be split into a generic Github interface
class attached to no repository, and a subclass which is.  This will
happen later when we build the infrastructure to work with Gitpan
as a whole.