Make using BackPAN::Index safer across forks.
This lets Gitpan run in parallel, if it's careful.

DBD::SQLite does not play well across forks.  DBIx::Class does not
shield us from this.  Since BackPAN::Index is the only thing
touching a database, the problem is mitigated if we get a new
instance of BackPAN::Index after forking.