Combine LEOCHARRE-DEBUG into LEOCHARRE-Debug.
This is a tough one, because they are two different modules, not just renames.
LEOCHARRE::Debug is not just a recasing of LEOCHARRE::DEBUG.  LEOCHARRE::DEBUG
has more content and releases, but is listed as deprecated in favor of
LEOCHARRE::Debug which is mostly a stub.

I decided it is worth preserving the history of LEOCHARRE-DEBUG, yet
LEOCHARRE-Debug is the way forward.  So shoving LEOCHARRE-DEBUG into
the LEOCHARRE-Debug repository is a compromise.