Nagios-NRPE
This is a tough one.  Both Nagios-NRPE and Nagios-Nrpe are recent,
well developed modules.  There is no clear winner.

I chose to rename the newer one (who should have done the research to look
for a namespace clash) to append their CPANID so the history is in gitpan.

I informed both authors about the problem.
https://rt.cpan.org/Ticket/Display.html?id=99485