Clone with a backoff in case the repository doesn't exist.
This is the solution to the mysterious "Unexpected HTTP status code: 404"
we're getting from clone.  Try it a few times in case the Github API
lied to us and is still making the repository or in case of a
network hiccup.

Also improve the backoff error diagnostics for push and pull

For #90