Add Gitpan::Repo->prepare_for_commits()
This will get the repository ready to take commits on demand, rather
than doing it as soon as one uses the git accessor.  This does not
prepare the repository for pushing, that's another step.  This is
to avoid prematurely creating a Github repository which is
expensive and can get in the way if it winds up not being used.

If there's both a Git and Github repository already, the goal is to
make sure they have not diverged and to get the local Git repository
caught up with the Github repo.  "git pull --ff-only" takes care
of all of this.  If they have diverged, this is an error which a
human must resolve.