Add Gitpan::Github->branch_info
Get information about a branch from the Github API.  Useful to test the
state of the Github repository without having to clone it.

Also increase the default backoff to 4 times.  This is probably more useful
for testing than production as we're not usually working on the same
repository in parallel.