Test that import_releases() will not create an empty repository.
Some distributions have no importable releases.  This could lead to
creating a Github repository and then doing nothing with it.  This
can then lead to a distribution with the same name, different case,
to fail to import because it thinks there's a clashing repository.

Using reform/ReForm because that is a real pair of clashing
releases.  ReForm has no unskipped releases.

Gitpan::Dist->git now creates a Github repository on demand which
fixes the problem.  I put the check to bail out if there's no releases
earlier to avoid a bogus duplicate repository error.