new_repo() and new_dist() test functions to create Repo and Dist objs for testing
These make sure the repository or distribution in question doesn't clash
with an existing one (via a random name), that the Github repository has
been deleted (and wait until the API catches up), and to clean it up
when the test is over.

Some do not use this because we want the test repositories to stick around
for examination.

Also...
* Add Repo->releases which does not create a Git repository just to check
  there are no releases