Add Repo->sync_with_github
This syncs with Github if there is a Github repository, otherwise
does nothing.  This is the guts of prepare_for_push() if there is
a Github repo.

This is necessary to be done before certain operations which want to
information about the repository using the local git repository such
as releases().

Moved a lot of the prepare_for_push() tests into sync_with_github()
to avoid redundancy.