Convert change_repo_info() to Pithub.
It was untested because it's not used... and it didn't work!  I've
left it in place because it will be useful as a utility to change
all the repositories in the future.

I also learned about the client errors we should be looking for
that indicate a real problem.  I'll implement this next.
https://developer.github.com/v3/#client-errors