Truncate files which exceed Github's limit of 100 megs per file.
This at least lets us import the release.

There's only two distributions with this problem.

I decided to completely truncate the file, rather than truncate it at the limit,
to make it very obvious to the reader that the file has been truncated.

The file in question is not the real tarball, it's 50+ megs.  The 100+
meg file in question has been replaced with a 101 meg file of all
zeroes so it compresses well.  This avoids putting a huge file into
the repository.

For #110